+++ lib/sakurraform/cli/plan.rb
@@ -26,10 +26,17 @@ module SakurraForm
       col_networks.resources.each do |net|
         unless net.resource_id
           net.resource_id = name_plus_uuid(net.name)
-          options = net.configuration.first.merge({'name' => net.resource_id})
-          say("Create new network #{net.name}")
-          router = network.routers.create(options)
-          switch = network.switches.find {|s| s.internet['ID'] == router.id}
+          options = net.configuration.merge({'name' => net.resource_id})
+          say("Create new #{net.mode} #{net.name}")
+          case net.mode
+          when 'router'
+            router = network.routers.create(options)
+            switch = network.switches.find {|s| s.internet['ID'] == router.id}
+          when 'switch'
+            switch = network.switches.create(options)
+          else
+            raise "Not supported mode #{net.mode}..."
+          end
           create_file "state/network/#{net.resource_id}.yml", switch.all_attributes.to_yaml
         else
           say("#{net.name} already available as #{net.resource_id}")
@@ -47,8 +54,8 @@ module SakurraForm
       col_servers.resources.each do |sv|
         unless sv.resource_id
           sv.resource_id = name_plus_uuid(sv.name)
-          switch_id = resolve_sakura_id_by_combined(sv.configuration.first["switch"])
-          options = sv.configuration.first.merge(
+          switch_id = resolve_sakura_id_by_combined(sv.configuration["switch"])
+          options = sv.configuration.merge(
             {
               "name"   => sv.resource_id,
               "switch" => switch_id
@@ -58,7 +65,7 @@ module SakurraForm
           server = compute.servers.create(options)
 
           sv_network = (col_networks.resources.find {|n| n.cached_state[:id] == switch_id}).cached_state[:subnets].first
-          sv_ipaddress = get_offset_address(sv_network["DefaultRoute"], sv.configuration.first["meta"]["network_offset"])
+          sv_ipaddress = get_offset_address(sv_network["DefaultRoute"], sv.configuration["meta"]["network_offset"])
           subnet ={
             :ipaddress => sv_ipaddress,
             :networkmasklen => sv_network["NetworkMaskLen"],
@@ -68,6 +75,22 @@ module SakurraForm
           disk_id = server.disks.first['ID']
           say("Associate #{sv_ipaddress} to #{sv.name}")
           volume.associate_ip_to_disk(disk_id, subnet)
+          say("Carve #{sv.name} as Hostname...")
+          volume.carve_hostname_on_disk(disk_id, sv.name)
+
+          ## Regist Interfaces
+          if sv.configuration["interfaces"]
+            ifs = sv.configuration["interfaces"]
+            c_switches = ifs.map do |target_sw|
+              say("Creating interface connected to #{target_sw}...")
+              resolve_sakura_id_by_combined(target_sw)
+            end
+            c_switches.map do |t_sw|
+              new_if = network.interfaces.regist_onto_server(server.id)
+              network.interfaces.connect_to_switch(new_if.id, t_sw)
+            end
+          end
+
           server.boot
           create_file "state/server/#{sv.resource_id}.yml", server.all_attributes.to_yaml
         else
@@ -124,7 +147,7 @@ module SakurraForm
       col_networks.resources.each do |net|
         if net.remote_state
           # puts net.remote_state
-          if net.remote_state[:internet].any?
+          if net.remote_state[:internet] && net.remote_state[:internet].any?
             say("Deleting Router #{net.resource_id} ...")
             network.delete_router(net.remote_state[:internet]["ID"])
           else