+++ lib/sakurraform/resource/base.rb
@@ -8,7 +8,7 @@ module SakurraForm
         type = self.class.to_s.split('::').last.downcase
         @name = name
         @resource_id = resolve_id_by_name(type, name)
-        @configuration = SakurraForm::Plans[type.to_sym].select {|a| a.name == name}
+        @configuration = (SakurraForm::Plans[type.to_sym].select {|a| a.name == name}).first
         @cached_state = @resource_id  ? collect_cached_state(type) : nil
         enable_remote = false unless @cached_state
         @remote_state = enable_remote ? collect_remote_state : nil